requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==2.3.2
scipy==1.14.1